+++ packages/core/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./doctor.js";
 export * from "./repo-layout.js";
-export * from "./lib/env.js";
+export * from "./lib/deploy-creds.js";